+++ agents-cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inkeep/agents-cli
 
+## 0.63.3
+
+### Patch Changes
+
+- @inkeep/agents-manage-ui@0.63.3
+- @inkeep/agents-core@0.63.3
+- @inkeep/agents-sdk@0.63.3
+
 ## 0.63.2
 
 ### Patch Changes